+++ packages/qb/pqb/src/sql/common.ts
@@ -34,8 +34,8 @@ export const revealColumnToSql = (
     return `${quoted}.${q(
       getJoinedColumnName(data, shape, table, key, quoted === quotedAs) || key,
     )}`;
-  } else if (quotedAs) {
-    return `${quotedAs}.${q(shape[column]?.data.name || column)}`;
+  } else if (quotedAs && shape[column]) {
+    return `${quotedAs}.${q(shape[column].data.name || column)}`;
   } else {
     return q(shape[column]?.data.name || column);
   }